dataloader for deeponet for different trunk inputs
