+++ src/torchphysics/utils/data/__init__.py
@@ -1,1 +1,2 @@
-from .dataloader import PointsDataset, PointsDataLoader, DeepONetDataLoader
+from .dataloader import PointsDataset, PointsDataLoader
+from .deeponet_dataloader import DeepONetDataLoader
